Move claim and reset password logic to their own controller before getting rid of OrcidProfile
